requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
Flask-DebugToolbar==0.13.1
Flask-WTF==1.0.1
WTForms-Components==0.10.5
Flask-Login==0.6.1
Flask-Mailman==0.3.0
Flask-Nav==0.6
Flask-Caching==1.10.1
bcrypt==3.2.2
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
matplotlib==3.5.1
seaborn==0.11.2
aif360==0.4.0
ipywidgets==7.7.0
itables==0.4.6
progressbar2==4.0.0
celery[redis]==5.2.6
python-dotenv==0.20.0
itsdangerous==2.1.2
